aurascript/requirements.txt
# AuraScript — Production dependencies (pinned)
# Install with: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12    # Required for UploadFile / form parsing

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Google Vertex AI ──────────────────────────────────────────────────────────
google-cloud-aiplatform==1.74.0
google-auth==2.36.0

# ── Async I/O ─────────────────────────────────────────────────────────────────
aiofiles==24.1.0
aiohttp==3.11.10
httpx==0.28.1               # Async HTTP client + test transport

# ── File magic / MIME detection ───────────────────────────────────────────────
python-magic==0.4.27        # Requires libmagic1 on host (apt/brew)

# ── Audio processing ──────────────────────────────────────────────────────────
# ffmpeg and ffprobe must be installed separately on the host.
# No Python bindings needed; we shell out via asyncio.create_subprocess_exec.

# ── Structured logging ────────────────────────────────────────────────────────
structlog==24.4.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
